pyproject.toml
[tool.poetry]
name = "ai-box"
version = "0.1.0"
description = ""
authors = ["Alexander Averyanov <alex@averyan.ru>"]

[tool.poetry.dependencies]
python = ">=3.11"
fastapi = "^0.111"
uvicorn = "^0.20"
python-multipart = "^0.0.9"
pillow = "^10.4"
pydantic = "^2.8"

torchvision = { version = "^0.19.0", source = "torch" }
torch = {version = "^2.4.0+cpu", source = "torch"}
transformers = "^4.43"
sentencepiece = "^0.2.0"
sacremoses = "^0.1.1"

[[tool.poetry.source]]
name = "torch"
url = "https://download.pytorch.org/whl/cpu"
priority = "supplemental"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
